+++ lib/src/widgets/async_button/async_button_style_button.dart
@@ -256,7 +256,7 @@ extension AsyncButtonExtension on ButtonStyleButton {
   ///
   /// - If [config] is set, all other parameters are ignored.
   /// - If any icon is set, the respective builder is ignored.
-  /// - If any color is set, the respective themer is ignored.
+  /// - If any color is set, the respective theme is ignored.
   Widget asAsync({
     AsyncButtonConfig? config,
     Widget? loadingIcon,
@@ -268,9 +268,9 @@ extension AsyncButtonExtension on ButtonStyleButton {
     Color? loadingColor,
     Color? errorColor,
     Color? successColor,
-    AsyncThemer? loadingThemer,
-    AsyncThemer? errorThemer,
-    AsyncThemer? successThemer,
+    AsyncThemeGetter? loadingTheme,
+    AsyncThemeGetter? errorTheme,
+    AsyncThemeGetter? successTheme,
     bool? keepHeight,
     bool? keepWidth,
     bool? animateSize,
@@ -301,9 +301,9 @@ extension AsyncButtonExtension on ButtonStyleButton {
             successDuration: successDuration,
             styleDuration: styleDuration,
             styleCurve: styleCurve,
-            errorThemer: errorThemer,
-            loadingThemer: loadingThemer,
-            successThemer: successThemer,
+            errorTheme: errorTheme,
+            loadingTheme: loadingTheme,
+            successTheme: successTheme,
           ),
       configurator: (context) {
         final config = Async.of(context);