+++ lib/src/widgets/async_button/async_floating_action_button.dart
@@ -157,7 +157,7 @@ extension AsyncFloatingActionButtonExtension on FloatingActionButton {
   ///
   /// - If [config] is set, all other parameters are ignored.
   /// - If any icon widget is set, the respective builder is ignored.
-  /// - If any seed color is set, the respective themer is ignored.
+  /// - If any seed color is set, the respective theme is ignored.
   Widget asAsync({
     AsyncButtonConfig? config,
     Widget? loadingIcon,
@@ -169,9 +169,9 @@ extension AsyncFloatingActionButtonExtension on FloatingActionButton {
     Color? loadingColor,
     Color? errorColor,
     Color? successColor,
-    AsyncThemer? loadingThemer,
-    AsyncThemer? errorThemer,
-    AsyncThemer? successThemer,
+    AsyncThemeGetter? loadingTheme,
+    AsyncThemeGetter? errorTheme,
+    AsyncThemeGetter? successTheme,
     bool? keepHeight,
     bool? keepWidth,
     bool? animateSize,
@@ -205,9 +205,9 @@ extension AsyncFloatingActionButtonExtension on FloatingActionButton {
                 successDuration: successDuration,
                 styleDuration: styleDuration,
                 styleCurve: styleCurve,
-                errorThemer: errorThemer,
-                loadingThemer: loadingThemer,
-                successThemer: successThemer,
+                errorTheme: errorTheme,
+                loadingTheme: loadingTheme,
+                successTheme: successTheme,
               ),
           configurator: (ctx) => Async.of(ctx).floatingActionButtonConfig,
           onPressed: onPressed,